Make test not fail in teardown
`test_postgres_logical.py` creates a test database every time it runs, which gets deleted in teardown. It also expectedly throws an error if it cannot add any more replication slots. This can happen if `max_replication_slots=1` and the replication slot is being used by the dev environment.

When it does throw the "cannot add any more replication slots" error, we are still holding a connection to the temporary database. So, in teardown, when we try to drop the database, we get another error:
```
psycopg2.OperationalError: database "cdc_test_6673d912a4cd11e98136acde48001122" is being accessed by other users
DETAIL:  There is 1 other session using the database.
```
And then, we have a test database in postgres that never gets deleted.

With this PR, we close off all connections to the temporary test database before dropping the database.