fix(snapshot) Changes the type provided to the producer to fix the snapshot process. (#27)
This seemingly complex change is actually a bug fix.
Currently the snapshot process just does not work. Months ago we (ok I broke it) made changes to the Producer to embed a header that contains the table name in Kafka messages produced by the cdc producer. In order to do that I made the stream/Producer take a ReplicationMessage object instead of a bytes payload.
This was ok, for the normal cdc operation. But it broke the snapshot as the snapshot controller uses the consumer to send a simple bytes payload to Kafka. The test passed because it was mocking the logic too far from Kafka.

To fix that I introduced a StreamMessage to replace ReplicationMessage in the stream/producer interface.

stream message is a slightly abstract version of a message for the stream containing a payload (bytes) and a metadata mapping, which becomes the headers on Kafka. We could make it even more abstract but that would require more calls and more object instantiations (intermediate representation between stream and kafka) for every message processed. I would not add this additional loss in performance now considering how critical the message processing time is here.
Both control messages (for the snapshot) and ReplicationMessage (default cdc message) have a to_stream method that produces the StreamMessage above
the rest is just reshuffling the code to avoid circular dependencies and fix tests.