feat(snapshot) Provide a formatter configuration when taking a snapshot. (#28)
When we take a snapshot of a postgres table with cdc (to import it into clickhouse) dates are formatted as 2019-06-16 06:21:39+00. The process of transforming these dates into clickhouse dates (2019-06-16 06:21:39) is by far the bottleneck when ingesting the snapshot into Clickhouse. This bottleneck is specifically impactful on tables like groupedmessages_local that are huge.

We can largely avoid this overhead by just formatting the snapshot in a way that is good for clickhouse. Just piping the CSV into clickhouse is one order of magnitude faster.
At the same time I would not like to tightly couple the cdc producer with clickhouse.

So this PR adds more structure to the snapshot configuration so that we can configure a format for each column we query from postgres. This puts the coupling in the configuration we pass to cdc each time we extract a snapshot, instead of being in the cdc code itself.
The formatting logic is applied in the postgres query.