feat(snapshot) Compress table files after taking the snapshot (#29)
In certain cases, it is convenient to zip the table files after writing the snapshot and before using to ingest it somewhere else.
The main use case is when we need to ingest some tables contained in the snapshot via HTTP without the need of any pre-processing (we can do this with clickhouse for some tables like groupedmessages).
When that happens, ingeting a gzip file is a lot cheaper than a decompressed one. The difference in the ingestion time is more than enough to justify the time spent to gzip the file.

This allows the snapshot configuration to define whether a table has to be compressed or not.
The same snapshot can contain compressed and uncompressed table files.
In this current version compression happens for each table before closing the transaction. The transaction is not locking anything though, so this should not be a problem. We could optimize it by closing the transaction first if needed.