feat(docker): Adds a cloudbuild file to publish the docker image (#34)
This adds a new cloudbuild file to create the docker image and publish it on the public repo.

Specifically:

It does not remove the old cloudbuild so we can merge the new one for the public repo, have the first image published and, once that is verified to work properly remove the old one. During this process it is desirable to preserve the old file in a functioning state to be able to apply a fix and do a deployment if needed.
It adds a docker compose file to let cloudbuild run an integration test before publishing the image.